+++ constants/index.ts
@@ -9,3 +9,5 @@ export * from './imagesMapHafs';
 export * from './slides';
 
 export * from './readingPositionBannerHeights';
+
+export * from './riwayaOptions';